Updated numbering, common postconditions and description
